Refactor ProposerDuties
